Rename Event to RepositoryEvent
