+++ Radar/Radar.csproj
@@ -63,8 +63,8 @@
     <Compile Include="Clients\ForkConfiguration.cs" />
     <Compile Include="Clients\RepositoryClient.cs" />
     <Compile Include="Clients\RepositoryClientConfiguration.cs" />
-    <Compile Include="Event.cs" />
     <Compile Include="Clients\DummyEvent.cs" />
+    <Compile Include="Tracking\RepositoryEvent.cs" />
     <Compile Include="Identity.cs" />
     <Compile Include="IEvent.cs" />
     <Compile Include="Images\ImageManagerConfiguration.cs" />
