+++ Radar/Tracking/RepositoryEvent.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 
-namespace Radar
+namespace Radar.Tracking
 {
-    public class Event : IEvent
+    public class RepositoryEvent : IEvent
     {
         public Identity Identity { get; internal set; }
         public DateTime Time { get; internal set; }
